perf(layout): u16 inner-node children + SIMD u16 child scan
Store ART inner-node child slot indices as u16 instead of u32. Slot
indices are 1-based into a MAX_SLOTS=10240 table, so 16 bits suffice.
This roughly halves the inner-node footprint:

  Node4:    24 ->  16 B
  Node16:   88 ->  56 B
  Node48:  456 -> 360 B
  Node256:1032 -> 520 B   (the hot direct-indexed node: -49.6%)

All sizes remain multiples of 8, so the bump allocator keeps 8-byte
body alignment. Header field offsets (offset_of(children)) are
unchanged; only element width + size_of change. For read-heavy
metadata workloads this means denser blobs (more nodes per 512 KB
frame), fewer cache lines per descent, and less I/O per cold read.

Replace the now-dead u32 children scan (find_next_nonzero_u32 and its
SSE2/AVX2/NEON helpers) with a real SIMD u16 scan:
  - find_next_nonzero_u16: AVX2 16-lane on x86, SSE2/NEON 8-lane,
    scalar tail. Lane helpers return an even-bit mask so
    trailing_zeros()/2 recovers the first non-zero lane. All loads
    are unaligned, so it is sound on both 8-byte-aligned blob bodies
    and 2-byte-aligned stack copies of node structs.
  - Differential property test (find_next_nonzero_u16_matches_scalar)
    over lengths 4/16/48/256 and every start offset.

The internal walker representation stays u32; conversion happens at
the node-body read/write boundary (the type checker enforces every
narrowing site), so the change is mechanical and fully checked.

NOTE: this is an on-disk format break (node sizes changed). Pre-1.0,
existing .holt files are not migrated; recreate test/dev data.

Tests: bumped multi-blob test key counts (snapshot.rs, tree_smoke.rs)
from 2000 to 5000 — denser packing means the old volume no longer
forces a multi-blob tree. Full suite green (lib + integration +
proptest, 0 failures).

+++ src/engine/walker/migrate.rs
@@ -443,11 +443,11 @@ fn clone_node4(
         }
         pack_inner_node(dst, &survivors)
     } else {
-        let mut new_children = [0u32; 4];
+        let mut new_children = [0u16; 4];
         for (i, slot) in new_children.iter_mut().enumerate().take(count) {
             let cloned = clone_subtree(src, dst, src_n.children[i] as u16, false)?
                 .expect("preserve mode never returns None");
-            *slot = u32::from(cloned);
+            *slot = cloned;
         }
         let out = dst.alloc_node(NodeType::Node4)?;
         let mut new_n = Node4::empty();
@@ -476,11 +476,11 @@ fn clone_node16(
         }
         pack_inner_node(dst, &survivors)
     } else {
-        let mut new_children = [0u32; 16];
+        let mut new_children = [0u16; 16];
         for (i, slot) in new_children.iter_mut().enumerate().take(count) {
             let cloned = clone_subtree(src, dst, src_n.children[i] as u16, false)?
                 .expect("preserve mode never returns None");
-            *slot = u32::from(cloned);
+            *slot = cloned;
         }
         let out = dst.alloc_node(NodeType::Node16)?;
         let mut new_n = Node16::empty();
@@ -518,12 +518,12 @@ fn clone_node48(
         }
         pack_inner_node(dst, &survivors)
     } else {
-        let mut new_children = [0u32; 48];
+        let mut new_children = [0u16; 48];
         for (i, slot) in new_children.iter_mut().enumerate() {
             if src_n.children[i] != 0 {
                 let cloned = clone_subtree(src, dst, src_n.children[i] as u16, false)?
                     .expect("preserve mode never returns None");
-                *slot = u32::from(cloned);
+                *slot = cloned;
             }
         }
         let out = dst.alloc_node(NodeType::Node48)?;
@@ -555,12 +555,12 @@ fn clone_node256(
         }
         pack_inner_node(dst, &survivors)
     } else {
-        let mut new_children = [0u32; 256];
+        let mut new_children = [0u16; 256];
         for (i, slot) in new_children.iter_mut().enumerate() {
             if src_n.children[i] != 0 {
                 let cloned = clone_subtree(src, dst, src_n.children[i] as u16, false)?
                     .expect("preserve mode never returns None");
-                *slot = u32::from(cloned);
+                *slot = cloned;
             }
         }
         let out = dst.alloc_node(NodeType::Node256)?;
@@ -614,7 +614,7 @@ fn pack_inner_node(dst: &mut BlobFrame<'_>, survivors: &[(u8, u32)]) -> Result<O
             n.count = survivors.len() as u8;
             for (i, &(b, c)) in survivors.iter().enumerate() {
                 n.keys[i] = b;
-                n.children[i] = c;
+                n.children[i] = c as u16;
             }
             write_struct_to_slot(dst, out.slot, &n)?;
             Ok(Some(out.slot))
@@ -625,7 +625,7 @@ fn pack_inner_node(dst: &mut BlobFrame<'_>, survivors: &[(u8, u32)]) -> Result<O
             n.count = survivors.len() as u8;
             for (i, &(b, c)) in survivors.iter().enumerate() {
                 n.keys[i] = b;
-                n.children[i] = c;
+                n.children[i] = c as u16;
             }
             write_struct_to_slot(dst, out.slot, &n)?;
             Ok(Some(out.slot))
@@ -635,7 +635,7 @@ fn pack_inner_node(dst: &mut BlobFrame<'_>, survivors: &[(u8, u32)]) -> Result<O
             let mut n = Node48::empty();
             n.count = survivors.len() as u8;
             for (ci, &(b, c)) in survivors.iter().enumerate() {
-                n.children[ci] = c;
+                n.children[ci] = c as u16;
                 n.index[b as usize] = (ci as u8) + 1;
             }
             write_struct_to_slot(dst, out.slot, &n)?;
@@ -649,7 +649,7 @@ fn pack_inner_node(dst: &mut BlobFrame<'_>, survivors: &[(u8, u32)]) -> Result<O
             // the count field is informational.
             n.count = survivors.len() as u8;
             for &(b, c) in survivors {
-                n.children[b as usize] = c;
+                n.children[b as usize] = c as u16;
             }
             write_struct_to_slot(dst, out.slot, &n)?;
             Ok(Some(out.slot))
